pyproject.toml
[project]
name = "logios-brain"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.1",
    "psycopg>=3.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "supabase>=2.7.4",
    "qdrant-client>=1.10.1",
    "neo4j>=5.23.1",
    "google-genai>=0.1.0",
    "httpx>=0.27.2",
    "pydantic>=2.8.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
